Remove UpdateDataInFile function from client management
The function `UpdateDataInFile` has been completely removed. This function was responsible for updating client data in a file by iterating through a vector of clients and writing either the new client data or existing client data to the specified file. Its removal indicates a change in the approach to managing or storing client data in the application.